Updated version to v5.1.2

+++ generationPython.sh
@@ -13,6 +13,7 @@ check_format_1="^[0-1][0-9]:[0-5][0-9]:[0-5][0-9]" # Checks times from 00:00:00 
 check_format_2="^[2][0-3]:[0-5][0-9]:[0-5][0-9]" # Checks times from 20:00:00 to 23:59:59
 
 time_now=$(date +%H%M%S) # Time at this moment
+logger=1 # Variable to show waiting message
 
 # Checks file content (scheduler.cfg)
 if [[ ! -z "$content_scheduling" && -s $scheduler_file ]]
@@ -73,13 +74,15 @@ latitude_code=$(head -n 10 $parameter_file | tail -n 1 | grep -o '^[^#]*' | grep
 altitude=$(head -n 11 $parameter_file | tail -n 1 | grep -o '^[^#]*' | grep -o '[^altitude=]*')
 object=$(head -n 12 $parameter_file | tail -n 1 | grep -o '^[^#]*' | grep -o '[^object=].*')
 content=$(head -n 13 $parameter_file | tail -n 1 | grep -o '^[^#]*' | grep -o '[^content=].*')
+control_external_generation=$(head -n 14 $parameter_file | tail -n 1 | grep -o '^[^#]*' | grep -o '[^control_external_generation=].*')
 
 # Check the content of the variables
 if [[ -z $freq_min || -z $freq_max || -z $gain ||
       -z $gen_mode   || -z $station_name || -z $focus_code || 
       -z $longitude  || -z $longitude_code ||
       -z $latitude   || -z $latitude_code  || -z $altitude ||
-      -z $object   || -z $content ]]
+      -z $object   || -z $content ||
+      -z $control_external_generation ]]
 then
     echo "File parameters are empty"
     echo "...Exiting..."
@@ -88,39 +91,43 @@ else
     # Check gen mode and if gen_mode == 0, then scheduling.cfg will be read and sw will be executed
     if [ $gen_mode -eq 0 ]
     then        
-        while read schedule_time
+        while [ 1 ]
         do
-            # Add time to have more margin
-            schedule_time_future=$(date -d "$schedule_time  + 1 seconds
-                                                            + 1 seconds
-                                                            + 1 seconds
-                                                            + 1 seconds
-                                                            + 1 seconds
-                                                            + 1 seconds
-                                                            + 1 seconds
-                                                            + 1 seconds" +"%H%M%S")
-            while [ $time_now -lt $schedule_time_future ]
-            do
-                time_now=$(date +%H%M%S) 
-            done
+            control_external_generation=$(head -n 14 $parameter_file | tail -n 1 | grep -o '^[^#]*' | grep -o '[^control_external_generation=].*')
+            if [ $control_external_generation -eq 1 ]
+            then
+                # execute generation with python
+                cd pythonScripts/
+                python3 generationFits.py $station_name $focus_code $latitude $latitude_code $longitude $longitude_code $altitude $object $content
+                rm samples.txt
+                rm times.txt
+                rm frequencies.txt
+                rm header_times.txt
 
-            # execute generation with python
-            cd pythonScripts/
-            python3 generationFits.py $station_name $focus_code $latitude $latitude_code $longitude $longitude_code $altitude $object $content
-            rm samples.txt
-            rm times.txt
-            rm frequencies.txt
-            rm header_times.txt
-
-            cd ..       
-            # sed -i 's\control_external_generation=1\control_external_generation=0\' $parameter_file # TODO: Checks
+                cd ..       
             
-            mv pythonScripts/*.fit Result/LastResult
-            mv pythonScripts/*_logs.txt Result/LastResult
+                mv pythonScripts/*.fit Result/LastResult
+                mv pythonScripts/*_logs.txt Result/LastResult
+                
+                echo "...Fits file generated..."
+                
+                # Disable control flag to not execute python script again
+                sed -i 's\control_external_generation=1\control_external_generation=0\' $parameter_file
+                logger=1
+            else
 
-        done < $scheduler_file
+                if [ $logger -eq 1 ]
+                then
+                    echo "...Waiting until sweeping is done..."
+                    logger=0                
+                fi
+
+            fi 
+        done
+
     else 
-        echo "C generation. You dont need to execute this"
+        echo "C generation. You dont need to execute this."
+        echo "If you want to execute python generation change gen_mode to 0 at config.cfg."
         echo "...Exiting..."
         exit 0
